refactor(searchbar): move to script setup
