+++ src/packages/__VUE/searchbar/doc.md
@@ -91,6 +91,19 @@ app.use(Searchbar)
 | rightin | 输入框内 右 icon |
 | rightout | 输入框外 右 icon |
 
+### 类型定义 version
+
+组件导出以下类型定义：
+
+```js
+import type {
+  SearchbarProps,
+  SearchbarInstance,
+  SearchbarInputAlign,
+  SearchbarShape,
+} from '@nutui/nutui';
+```
+
 ## 主题定制
 
 ### 样式变量
